Added more terminals
Added more terminals to search for in test_terminals, 1 exclusively for wayland, and 7 for X11. There is a problem for zutty where the default font is set to 9x18, and it may not have a universal default that will be on every system.

One solution would be to supply a font for zutty, but that would add something most people wouldn't use.